v1.19.0 — kuantitas pecahan (jual sebagian), skema tetap v11
Fitur: barang yang satuannya bukan bijian (kg, liter, ikat, gram, ml) boleh
dijual sebagian — ¼ kg gula, ½ liter minyak.

Skema sengaja TIDAK dinaikkan (tetap v11):
- SQLite memakai tipe kolom sebagai affinity, bukan batasan: di kolom
  `quantity INTEGER` nilai `2` disimpan sebagai integer dan `0,5` sebagai real.
  Kelima kolom kuantitas (`products.stock`, `products.min_stock`,
  `sale_items.quantity`, `sales.total_items`, `stock_movements.quantity`)
  karena itu sudah menampung pecahan apa adanya.
- Rencana lama di roadmap ("ubah `quantity INTEGER` menjadi `REAL`, skema
  versi 12") keliru dan berbahaya: SQLite tidak punya `ALTER COLUMN`, jadi
  satu-satunya cara adalah menulis ulang tabel `products`/`sales`/`sale_items`
  — membongkar tabel berisi seluruh riwayat penjualan pelanggan.
- Alasannya ditulis di doc block `database_helper.dart`, dan pemeriksa
  menjaga `_dbVersion` tetap 11 serta tidak adanya `_upgradeV12`.

Kelas kesalahan yang tidak bisa dilihat pemeriksa mana pun — `as int` pada
kolom kuantitas meledak saat nilainya 0,5, dan `as double` meledak saat
nilainya 2. Keduanya galat RUNTIME di HP pelanggan, bukan saat build. Karena
itu pembacaannya dipusatkan:
- `lib/core/utils/angka.dart` (baru): `Angka.jumlah` untuk kuantitas,
  `Angka.jumlahAtau` untuk kolom bercadangan, `Angka.uang` untuk uang
  (`.round()`, bukan `.toInt()` — memotong 0,5 rupiah melaporkan uang lebih
  kecil daripada yang benar-benar masuk).
- `lib/core/utils/satuan.dart` (baru): `bolehPecahan` diturunkan dari satuan
  (`satuanBulat`), bukan disimpan per barang — jadi tidak perlu kolom baru.
  Satu tempat dipakai keranjang, tombol cepat kasir, dan formulir produk.
- `Formatters.jumlah`: "2" tetap "2" (bukan "2,00"), "0.5" menjadi "0,5".

Uang tetap bilangan bulat:
- Sepuluh rumus yang mengalikan kuantitas dengan harga sekarang `.round()`.
- Rumus yang menjadi dasar seluruh laporan TIDAK berubah:
  `totalAmount = hargaBarang - potonganNota`,
  `totalProfit = labaBarang - potonganNota`.
- Laba baris tetap dari harga SETELAH potongan (`subtotal - modal × jumlah`);
  memakai harga label akan melaporkan laba lebih besar daripada uang masuk.

Layar kasir:
- Baris keranjang bisa diketuk untuk mengubah jumlah; dialognya menyediakan
  tombol cepat ¼ dan ½ plus isian manual yang menerima koma maupun titik.
- `CartNotifier.setQuantity()` baru, dan keempat mutasi keranjang memakai
  `Satuan.batasi` (tidak bisa melebihi stok, tidak bisa menyisakan nol) serta
  tetap membawa `discount: item.discount`.
- Petunjuk "Ketuk angka jumlah untuk mengisi ¼ atau ½" hanya muncul kalau ada
  barang yang satuannya boleh dipecah.

Ikut diperbaiki: README memuat bagian "Catatan ucapan struk dan grafik
penjualan" DUA kali — skrip sisip v1.18.0 tidak idempoten ("insert sebelum
jangkar" tidak mengubah jangkarnya, jadi run kedua menyisipkan blok yang sama
lagi). Duplikatnya dihapus.

Pemeriksa: check_phase_e 776/0 (naik dari 686), check_dart 86/0,
uji_redirect 38/0. `negatif_27.py` (81 mutasi + 4 hijau) UJI-NEGATIF LULUS.
Gate penuh LULUS.

Pra-terbang (tanpa argumen) menemukan 8 pola uji-negatif LAMA yang rusak
karena perubahan bentuk baris di rilis ini — 5 di `negatif_21.py`, 3 di
`negatif_22.py`. Semuanya diperbaiki dengan menunjuk ulang polanya, tidak ada
mutasi yang dihapus.

+++ lib/providers/product_provider.dart
@@ -92,7 +92,7 @@ class ProductNotifier extends StateNotifier<ProductState> {
     await loadProducts();
   }
 
-  Future<void> restock(int productId, int quantity) async {
+  Future<void> restock(int productId, double quantity) async {
     await _repo.addStock(productId, quantity);
     await loadProducts();
   }
@@ -101,7 +101,7 @@ class ProductNotifier extends StateNotifier<ProductState> {
   /// kalau belanjanya belum dibayar penuh.
   Future<void> restockProduct({
     required ProductModel product,
-    required int quantity,
+    required double quantity,
     int? costPerUnit,
     int paidNow = 0,
     String? supplierName,
